+++ client/src/features/search/components/tabs/SearchAll.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
 import { faker } from '@faker-js/faker';
-import Heading2 from '../../../ui/Typography/Heading2.tsx';
-import SortedTable from '../../../ui/Table/custom/SortedTable/SortedTable.tsx';
-import CardsList from '../../../ui/CardsList/CardsList.tsx';
+import Heading2 from '../../../../ui/Typography/Heading2.tsx';
+import SortedTable from '../../../../ui/Table/custom/SortedTable/SortedTable.tsx';
+import CardsList from '../../../../ui/CardsList/CardsList.tsx';
 
 interface CardItem {
   img: string;
