+++ client/src/features/search/components/tabs/SearchPlaylists.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { faker } from '@faker-js/faker';
-import CardsList from '../../../ui/CardsList/CardsList.tsx';
+import CardsList from '../../../../ui/CardsList/CardsList.tsx';
 
 interface CardItem {
   img: string;
