+++ client/src/layout/Nav/components/CenterNav.jsx
@@ -2,7 +2,7 @@ import styles from '../Nav.module.scss';
 import NavTooltip from './NavTooltip.jsx';
 import NavButton from './NavButton.jsx';
 import { RiHome4Line } from 'react-icons/ri';
-import NavForm from './NavForm.jsx';
+import NavForm from '../../../features/search/components/input/NavForm.jsx';
 
 const CenterNav = () => {
   return (
